requirements.txt
eth-brownie
black